Bug 1239072 - Refactored SettingsTableViewController into a subclass-able component. Moved App Settings into separate controller and created new Authentication settings VC.
