+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -2349,7 +2349,7 @@ extension BrowserViewController: IntroViewControllerDelegate {
         // Show the settings page if we have already signed in. If we haven't then show the signin page
         let vcToPresent: UIViewController
         if profile.hasAccount() {
-            let settingsTableViewController = SettingsTableViewController()
+            let settingsTableViewController = AppSettingsTableViewController()
             settingsTableViewController.profile = profile
             settingsTableViewController.tabManager = tabManager
             vcToPresent = settingsTableViewController
